Wishlist hook files refactor.
