--- packages/peregrine/lib/talons/Wishlist/WishlistButton/wishlistButton.gql.js
@@ -1,21 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const ADD_TO_WISHLIST = gql`
-    mutation addProductToWishlist(
-        $wishlistId: ID!
-        $itemOptions: WishlistItemInput!
-    ) {
-        addProductsToWishlist(
-            wishlistId: $wishlistId
-            wishlistItems: [$itemOptions]
-        ) {
-            user_errors {
-                code
-                message
-            }
-        }
-    }
-`;
-export default {
-    addProductToWishlistMutation: ADD_TO_WISHLIST
-};
